chore: deleta componentes inutilizados e atualiza importação
Foram excluídos vários componentes não utilizados relacionados ao painel, incluindo CardsHome, CardsRents, CardsSale, PropertyNumber, PropertySold, Section, UsersPerMonth e vários componentes de cartão de aluguel/venda. O LineGraph também foi movido para um novo local e sua importação na página do painel foi atualizada. Essa limpeza reduz a desorganização da base de código e simplifica a implementação do painel.
