--- front-end/src/components/dash/PropertySold.js
@@ -1,92 +1,0 @@
-"use client";
-
-import { useState, useEffect } from "react";
-import { Doughnut } from "react-chartjs-2";
-import { Chart as ChartJS, ArcElement, Tooltip, Legend } from "chart.js";
-import { Row, Col } from "antd";
-
-ChartJS.register(ArcElement, Tooltip, Legend);
-
-export default function PropertySold() {
-  const [chartData, setChartData] = useState(null);
-
-  // Puxando dados de uma API real, mas a API tá fora do ar no momento ↓
-  // useEffect(() => {
-  //   fetch("https://6a3f34f1-cba4-496f-bbb1-ab1de57caebb.mock.pstmn.io/regioes") // URL da sua API
-  //     .then((res) => res.json())
-  //     .then((data) => {
-  //       // aqui você ajusta o formato para o Chart.js
-  //       setChartData({
-  //         labels: data.map((item) => item.nome), // exemplo: "Região Sul"
-  //         datasets: [
-  //           {
-  //             data: data.map((item) => item.valor), // exemplo: 25
-  //             backgroundColor: [
-  //               "#243B7B",
-  //               "#F39C12",
-  //               "#B8AEBF",
-  //             ],
-  //             borderWidth: 0,
-  //             cutout: "60%",
-  //           },
-  //         ],
-  //       });
-  //     });
-  // }, []);
-
-  // Dados estáticos para demonstração  ↓
-  const data = {
-    labels: ["Apartamentos", "Casas", "Terrenos"],
-    datasets: [
-      {
-        data: [45, 25, 15],
-        backgroundColor: ["#243B7B", "#F39C12", "#B8AEBF"], // cores
-        borderWidth: 0,
-        cutout: "0%", // transforma em donut (se fosse 0%, seria uma pizza cheia)
-      },
-    ],
-  };
-
-  const options = {
-    plugins: {
-      legend: {
-        position: "right",
-        labels: {
-          usePointStyle: false,
-          boxHeight: 18,
-
-          color: "black",
-          boxWidth: 18,
-        },
-      },
-    },
-    maintainAspectRatio: false,
-  };
-
-  return (
-    <Row className="py-4">
-      <Col className="gutter-row" span={24}>
-        <div className="group h-[300px] !w-full flex items-center rounded-xl px-10 !bg-[#EEF0F9] !shadow-md">
-          <div className="grid grid-col content-evenly w-full h-full">
-            <span className="text-lg md:text-2xl font-bold lg:text-center text-[var(--primary)] ">
-              Proporção de imóveis vendidos
-            </span>
-
-            <div className="items-center justify-items-center w-full h-full">
-              <div className="w-fit h-fit">
-                <Doughnut data={data} options={options} />
-
-                {/* Puxando dados de uma API real, mas a API tá fora do ar no momento ↓ */}
-                {/* {chartData ? (
-              <Doughnut data={chartData} options={options} />
-            ) : (
-              <p>Carregando...</p>
-            )} */}
-              </div>
-            </div>
-          </div>
-        </div>
-      </Col>
-    </Row>
-  );
-}
